requirements.txt
requests == 2.31.0
Flask == 3.0.0
Flask-Bootstrap4 == 4.0.2
Flask-WTF == 1.2.1